% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{childANOVA2F}
\alias{childANOVA2F}
\title{Run the two-way ANOVA child.}
\usage{
childANOVA2F(dat, factors = c(1, 2), resp = 3,
  factor.names = names(dat[, factors]), resp.name = names(dat[, resp,
  drop = FALSE]), factor1.ref = levels(dat[, factors[1]])[1],
  factor2.ref = levels(dat[, factors[2]])[1], title = "title")
}
\arguments{
\item{dat}{dataframe}

\item{factors}{the column names or indices of the factors}

\item{resp}{the column name or index of the response}

\item{factor.names}{the names of the factor for the report
(default value is the column names)}

\item{resp.name}{the name of the response for the report
(default value is the column name)}

\item{factor1.ref}{the reference level of factor1}

\item{factor2.ref}{the reference level of factor2}

\item{title}{study title}
}
\description{
Run knitr child for two-way ANOVA; to be used as inline code in
a knitr document.
}
