% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{format_ftable}
\alias{format_ftable}
\title{Format a table of type \code{ftable}}
\usage{
format_ftable(x, digits = getOption("digits"))
}
\arguments{
\item{x}{a table of type \code{\link{ftable}}}

\item{digits}{number of digits to print}
}
\value{
A table which can be used in \code{kable}.
}
\description{
Format a table of type \code{ftable} for HTML printing.
}
\note{
This function is based on \code{R2HTML:::HTML.ftable}
}
