% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{SimDataAV2}
\alias{SimDataAV2}
\title{Simulate a two-way ANOVA design with random effects}
\usage{
SimDataAV2(I, J, Kmin, Kmax, p = NULL, mu = 0, sigmaP = 1,
  sigmaO = 1, sigmaPO = 1, sigmaE = 1, factor.names = c("Operator",
  "Part"), resp.name = "y", keep.intermediate = FALSE)
}
\arguments{
\item{I}{number of levels of first factor}

\item{J}{number of leveles of second factor}

\item{Kmin}{minimum number of repeats}

\item{Kmax}{maximum number of repeats}

\item{p}{numeric vector giving the probabilities to sample between
\code{Kmin} and \code{Kmax}}

\item{sigmaP}{standard deviation of first factor}

\item{sigmaO}{standard deviation of second factor}

\item{sigmaPO}{standard deviation of interaction}

\item{sigmaE}{residual standard deviation}

\item{factor.names}{names of the two factors}

\item{resp.name}{name of the response}

\item{keep.intermediate}{keep intermediate calculations in the output}
}
\value{
A dataframe.
}
\description{
Simulate a two-way ANOVA dataset with random effects.
}
\examples{
SimDataAV2(I=3, J=2, Kmin=0, Kmax=2, p=c(0.1,0.2))

}
